requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt pyproject.toml
#
aiofiles==25.1.0
    # via bio-engine (pyproject.toml)
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via starlette
asttokens==3.0.1
    # via stack-data
attrs==25.4.0
    # via
    #   bioutils
    #   hgvs
biocommons-seqrepo==0.6.11
    # via hgvs
biopython==1.86
    # via bio-engine (pyproject.toml)
bioutils==0.6.1
    # via
    #   bio-engine (pyproject.toml)
    #   biocommons-seqrepo
    #   hgvs
certifi==2026.1.4
    # via requests
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via uvicorn
coloredlogs==15.0.1
    # via biocommons-seqrepo
configparser==7.2.0
    # via hgvs
decorator==5.2.1
    # via ipython
exceptiongroup==1.3.1
    # via
    #   anyio
    #   ipython
executing==2.2.1
    # via stack-data
fastapi==0.131.0
    # via bio-engine (pyproject.toml)
h11==0.16.0
    # via uvicorn
hgvs==1.5.6
    # via bio-engine (pyproject.toml)
humanfriendly==10.0
    # via coloredlogs
idna==3.11
    # via
    #   anyio
    #   requests
importlib-metadata==8.7.1
    # via yoyo-migrations
importlib-resources==6.5.2
    # via hgvs
ipython==8.38.0
    # via
    #   biocommons-seqrepo
    #   hgvs
jedi==0.19.2
    # via ipython
matplotlib-inline==0.2.1
    # via ipython
numpy==2.2.6
    # via biopython
orjson==3.11.7
    # via bio-engine (pyproject.toml)
parsley==1.3
    # via hgvs
parso==0.8.6
    # via jedi
pexpect==4.9.0
    # via ipython
prompt-toolkit==3.0.52
    # via ipython
psycopg2==2.9.11
    # via hgvs
psycopg2-binary==2.9.11
    # via bio-engine (pyproject.toml)
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.3
    # via stack-data
pydantic==2.12.5
    # via
    #   bio-engine (pyproject.toml)
    #   fastapi
    #   pydantic-settings
pydantic-core==2.41.5
    # via pydantic
pydantic-settings==2.13.1
    # via bio-engine (pyproject.toml)
pygments==2.19.2
    # via ipython
pysam==0.23.3
    # via biocommons-seqrepo
python-dotenv==1.2.1
    # via pydantic-settings
requests==2.32.5
    # via
    #   biocommons-seqrepo
    #   bioutils
sqlparse==0.5.5
    # via yoyo-migrations
stack-data==0.6.3
    # via ipython
starlette==0.52.1
    # via fastapi
tabulate==0.9.0
    # via yoyo-migrations
tqdm==4.67.3
    # via biocommons-seqrepo
traitlets==5.14.3
    # via
    #   ipython
    #   matplotlib-inline
typing-extensions==4.15.0
    # via
    #   anyio
    #   biocommons-seqrepo
    #   exceptiongroup
    #   fastapi
    #   ipython
    #   pydantic
    #   pydantic-core
    #   starlette
    #   typing-inspection
    #   uvicorn
typing-inspection==0.4.2
    # via
    #   fastapi
    #   pydantic
    #   pydantic-settings
urllib3==2.6.3
    # via requests
uvicorn==0.41.0
    # via bio-engine (pyproject.toml)
wcwidth==0.6.0
    # via prompt-toolkit
yoyo-migrations==9.0.0
    # via biocommons-seqrepo
zipp==3.23.0
    # via importlib-metadata
